+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Two more regexes for a better detection of Lodash.
 - -a|--user-agent option to use a custom user-agent.
 - One more regex for a better detection of Handlebars.
+- One more regex to detect Highcharts from the comment at the top of
+  the file.
 
 ### Fixed